requirements.txt
python-chess>=1.10.0
berserk>=0.13.0
numpy
flask
requests